+++ R/ggplot.R
@@ -714,12 +714,16 @@ gg.SpatRaster <- function(data, ...) {
 #' The length of the vector mus correspond to the number of mesh vertices.
 #' The alternative name `colour` is also recognised.
 #' @param alpha A vector of scalar values setting the alpha value of the colors provided.
-#' @param edge.color Color of the mesh edges.
+#' @param edge.color Color of the regular mesh edges.
+#' @param edge.linewidth Line width for the regular mesh edges. Default 0.25
 #' @param interior If TRUE, plot the interior boundaries of the mesh.
-#' @param int.color Color used to plot the interior boundaries.
+#' @param int.color Color used to plot the interior constraint edges.
+#' @param int.linewidth Line width for the interior constraint edges. Default 0.5
 #' @param exterior If TRUE, plot the exterior boundaries of the mesh.
-#' @param ext.color Color used to plot the interior boundaries.
-#' @param crs A [CRS] object defining the coordinate system to project the mesh to before plotting.
+#' @param ext.color Color used to plot the exterior boundary edges.
+#' @param ext.linewidth Line width for the exterior boundary edges. Default 1
+#' @param crs A CRS object supported by [fm_transform()] defining the coordinate
+#' system to project the mesh to before plotting.
 #' @param nx Number of pixels in x direction (when plotting using the color parameter).
 #' @param ny Number of pixels in y direction (when plotting using the color parameter).
 #' @param mask A SpatialPolygon defining the region that is plotted.
@@ -733,10 +737,13 @@ gg.inla.mesh <- function(data,
                          color = NULL,
                          alpha = NULL,
                          edge.color = "grey",
+                         edge.linewidth = 0.25,
                          interior = TRUE,
                          int.color = "blue",
+                         int.linewidth = 0.5,
                          exterior = TRUE,
                          ext.color = "black",
+                         ext.linewidth = 1,
                          crs = NULL,
                          mask = NULL,
                          nx = 500, ny = 500,
@@ -774,7 +781,7 @@ gg.inla.mesh <- function(data,
 
     colnames(df) <- c("x", "y", "xend", "yend")
     mp <- ggplot2::aes(x = .data$x, y = .data$y, xend = .data$xend, yend = .data$yend)
-    msh <- ggplot2::geom_segment(data = df, mapping = mp, color = edge.color)
+    msh <- ggplot2::geom_segment(data = df, mapping = mp, color = edge.color, linewidth = edge.linewidth)
 
     # Outer boundary
     if (exterior) {
@@ -783,7 +790,7 @@ gg.inla.mesh <- function(data,
         data$loc[data$segm$bnd$idx[, 2], 1:2]
       )
       colnames(df) <- c("x", "y", "xend", "yend")
-      bnd <- ggplot2::geom_segment(data = df, mapping = mp, color = ext.color)
+      bnd <- ggplot2::geom_segment(data = df, mapping = mp, color = ext.color, linewidth = ext.linewidth)
     } else {
       bnd <- NULL
     }
@@ -798,7 +805,7 @@ gg.inla.mesh <- function(data,
       if (nrow(df) == 0) {
         int <- NULL
       } else {
-        int <- ggplot2::geom_segment(data = df, mapping = mp, color = int.color)
+        int <- ggplot2::geom_segment(data = df, mapping = mp, color = int.color, linewidth = int.linewidth)
       }
     } else {
       int <- NULL